refactor: 统一句柄宽度 —— PieToken 由 u64 收敛到 usize
问题：ABI 里唯一的 u64 句柄是 `PieToken`，而 `TaskId`/`TeamId`/`VirtAddr` 与
`mtu`/`bytes`/`millis` 全是 usize；`PieToken` 自己的 Wire 实现还要 `as usize` /
`as u64` 来回强转。在 riscv64 上两者同宽同布局，这个区别零收益；换到非 64 位目标
`as usize` 还会静默截断——不是保护，是隐患。

规矩（本次立下）：**句柄与索引一律 usize；只有语义上是固定宽度的数值才用 u64。**

改动（纯类型宽度，零语义）：
  - env：`PieToken(pub usize)`；`From`/`Wire`/四个 `FromPair` impl 里的 `as u64`
    全部消失；`dispatch` 的 `u64::from_le_bytes` → `usize::from_le_bytes`。
  - kernel：`Pie.token`、`next_pie_token()`、`AnyPie::token()`、
    `gate::{accord,revoke,release}`、`pole` 的 `mappings` key 与 `map/unmap/narrow`
    → usize；`envcall` 里 4 处 `token as usize` 直接删掉。
  - task：`mail.rs` 全部句柄签名、`handshake`（`send`/`recv` 载荷、`Quay`/`Pier`/
    `Referred`）、`directory`/`channel`/`service`，以及 7 个 bin 里装 token 的
    `AtomicU64` → `AtomicUsize`。
  - docs：`dispatch`/`ipc`/`root` 里 token、task id 的字段 `u64 LE` → `usize LE`，
    并修 `Binding { entry: u64 }` 那段样例。

保留 u64 的只有三类：时钟刻度（`now_ns`/`Uptime`）、位掩码（`Mprotect.flags`）、
协议「保留字段」的字节描述（无 Rust 类型）。

验证：fmt clean；build 0 error；clippy 0 error（仅余改动前既有警告）。
QEMU release + debug 双档：`spawn 1000` → 499500；`dir` → found + 枚举 echo；
`req hello` → "ifmmp…"；`hole` → "hi from shell…"；`clock` ok；`exit` → bye
→ `root: session over, shutting down` → `task: all tasks exited, system halted`。
**线上字节布局零变化**：riscv64 上 usize == u64，`to_le_bytes()` 仍 8 字节，
握手 `MTU = 9` 与 dispatch 字段偏移均不动。

+++ crates/env/src/dispatch.rs
@@ -10,9 +10,9 @@
 //! [0]      op      u8      1=Register 2=Unregister 3=Replace 4=Resolve 5=Enumerate 6=Connect
 //! [1..33]  name    [u8;32] Register/Unregister/Replace/Resolve/Connect：目标名字
 //!                          Enumerate：游标（全 0 = 从头开始）
-//! [33..41] entry   u64 LE  Register/Replace：入口门闩的目录侧 pie token
+//! [33..41] entry   usize LE  Register/Replace：入口门闩的目录侧 pie token
 //! [41..49] 保留    u64 LE  必须为 0
-//! [49..57] reply   u64 LE  调用方自带的回信 pie 的目录侧 token（0 = 无回复预期）
+//! [49..57] reply   usize LE  调用方自带的回信 pie 的目录侧 token（0 = 无回复预期）
 //! [57..64] 保留（0）
 //! ```
 //!
@@ -24,7 +24,7 @@
 //! ```text
 //! [0]      status  u8      0=Ok 1=Found 2=Connected 3=NotFound 4=Denied 5=Taken
 //! [1..33]  name    [u8;32] Found：这一页的名字 / Resolve 命中的名字
-//! [1..9]   entry   u64 LE  Connected：目录转授给调用方的入口门闩 token
+//! [1..9]   entry   usize LE  Connected：目录转授给调用方的入口门闩 token
 //! [9..17]  保留    u64 LE  必须为 0（原 owner 字段已删——服务 task id 由调用方
 //!                          用 `MailCall::Owned` 从 entry 的 `owner` 求得）
 //! ```
@@ -180,7 +180,7 @@ impl Request {
         {
             return Err(ProtocolError::Reserved);
         }
-        let entry = PieToken(u64::from_le_bytes(
+        let entry = PieToken(usize::from_le_bytes(
             m[ENTRY_AT..ENTRY_AT + 8].try_into().unwrap_or([0u8; 8]),
         ));
         let mut name_bytes = [0u8; NAME_LEN];
@@ -265,7 +265,7 @@ impl Reply {
                 })
             }
             STATUS_CONNECTED => {
-                let entry = PieToken(u64::from_le_bytes(
+                let entry = PieToken(usize::from_le_bytes(
                     m[REPLY_PAYLOAD_AT..REPLY_PAYLOAD_AT + 8]
                         .try_into()
                         .unwrap_or([0u8; 8]),
